+++ lib/features/design_your_plan/presentation/components/appointment_details.dart
